[3.1.0] Maintenance scripts emit and honor proper exit codes
git-svn-id: http://htmlpurifier.org/svnroot/htmlpurifier/trunk@1640 48356398-32a2-884e-a903-53898d9a118a

+++ maintenance/generate-includes.php
@@ -27,6 +27,7 @@ $exclude_files = array(
 // Determine what files need to be included:
 echo 'Scanning for files... ';
 $raw_files = $FS->globr('.', '*.php');
+if (!$raw_files) throw new Exception('Did not find any PHP source files');
 $files = array();
 foreach ($raw_files as $file) {
     $file = substr($file, 2); // rm leading './'
